+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-# ani-sdk
+# anime-sdk
 
 ## 1.0.0
 
 ### Major Changes
 
-- a266c7f: ani-sdk first version with 9 providers and http server
+- a266c7f: anime-sdk first version with 9 providers and http server